- address #10 by removing reactor-ipc interfaces and reactor-ipc dependency
